+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Version 2024.2
 *in development*
 
 * 🌟 Movies: add year and original language filter where possible for discover and search screens.
+* 🔧 Search: hide keyboard after submitting query. 
 
 #### 2024.2.3 🧪
 *2024-04-26*